chore(releases): fixing various test imports
